requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
enum-tools==0.12.0
frozenlist==1.5.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-cache==1.2.1
rich==13.9.3
shellingham==1.5.4
six==1.16.0
streamdeck==0.9.5
twitchAPI==4.3.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
url-normalize==1.4.3
yarl==1.16.0
